refactor: Move mnemonic API endpoints to a dedicated router file
